skill-judge-ai-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.4
python-docx==1.1.2
python-dotenv==1.0.1
groq==0.11.0
pydantic==2.9.2
python-multipart==0.0.9
